+++ crates/bevy_state/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "bevy_state"
+version = "0.14.0-dev"
+edition = "2021"
+description = "Bevy Engine's entity component system"
+homepage = "https://bevyengine.org"
+repository = "https://github.com/bevyengine/bevy"
+license = "MIT OR Apache-2.0"
+keywords = ["ecs", "game", "bevy"]
+categories = ["game-engines", "data-structures"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+default = ["bevy_reflect"]
+
+
+[dependencies]
+bevy_ecs = { path = "../bevy_ecs", version = "0.14.0-dev" }
+bevy_state_macros = { path = "macros", version = "0.14.0-dev" }
+bevy_utils = { path = "../bevy_utils", version = "0.14.0-dev" }
+bevy_reflect = { path = "../bevy_reflect", version = "0.14.0-dev", optional = true }
+
+[lints]
+workspace = true
+
+[package.metadata.docs.rs]
+rustdoc-args = ["-Zunstable-options", "--cfg", "docsrs"]
+all-features = true
