+++ crates/bevy_state/macros/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "bevy_state_macros"
+version = "0.14.0-dev"
+description = "Bevy ECS Macros"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+bevy_macro_utils = { path = "../../bevy_macro_utils", version = "0.14.0-dev" }
+
+syn = { version = "2.0", features = ["full"] }
+quote = "1.0"
+proc-macro2 = "1.0"
+
+[lints]
+workspace = true
+
+[package.metadata.docs.rs]
+rustdoc-args = ["-Zunstable-options", "--cfg", "docsrs"]
+all-features = true
