+++ crates/bevy_state/macros/src/lib.rs
@@ -1,0 +1,24 @@
+// FIXME(3492): remove once docs are ready
+#![allow(missing_docs)]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+
+extern crate proc_macro;
+
+mod states;
+
+use bevy_macro_utils::BevyManifest;
+use proc_macro::TokenStream;
+
+#[proc_macro_derive(States)]
+pub fn derive_states(input: TokenStream) -> TokenStream {
+    states::derive_states(input)
+}
+
+#[proc_macro_derive(SubStates, attributes(source))]
+pub fn derive_substates(input: TokenStream) -> TokenStream {
+    states::derive_substates(input)
+}
+
+pub(crate) fn bevy_state_path() -> syn::Path {
+    BevyManifest::default().get_path("bevy_state")
+}
