+++ examples/state/computed_states.rs
