+++ tools/publish.sh
@@ -11,6 +11,8 @@ crates=(
     bevy_reflect
     bevy_ecs/macros
     bevy_ecs
+    bevy_state/macros
+    bevy_state
     bevy_app
     bevy_time
     bevy_log
